Refactor article links in index view and update routes.rb
This commit refactors the article links in the index view of the ArticlesController. Instead of using a plain anchor tag, the link_to helper method is now used to generate the link. Additionally, the routes.rb file is updated to use the resources method for defining the routes for the articles.